Add comments / format wallet example
